refactor(batch-exports): move filter composition out of spmc
First of three PRs emptying `spmc.py`, the legacy pipeline module. Everything left
in it is live code that the current pipeline depends on, so the module gets taken
apart by concern rather than deleted.

`compose_filters_clause` and its two helpers move to `temporal/filters.py`. They
were a clean cut: the group needs nothing from the rest of `spmc.py`, and 16 of
its 17 imports were exclusive to it, so `spmc.py` no longer pulls in HogQL at all.

Placement is by caller, not by pipeline membership. Filter composition is used
from four different areas — `pipeline/`, `destinations/`, `temporal/` root and
`debug/` — so it sits flat in `temporal/` rather than under `pipeline/`, which
would have meant non-pipeline code reaching into the pipeline package.

Code moved verbatim; verified byte-identical against master.

Generated-By: PostHog Code
Task-Id: e2c0e744-189b-4ac0-a1f4-c999b0a0e071
